requirements.txt
boto3
yara-python
